Split handler & commands to increase coverage
